[dm-devel] [PATCH v2 20/26] block: Add bio_for_each_segment_all()
This is part of the immutable bvec prep work; bio_for_each_segment() is
going to have a different implementation so these need to be split
apart.

This change is also to better document the intent of code that's using
it - bio_for_each_segment_all() is only legal to use for code that owns
the bio.

Signed-off-by: Kent Overstreet <koverstreet google com>
CC: Jens Axboe <axboe kernel dk>